rename functions parser files
